expenses and dashboard enhancements

+++ database/migrations/2023_03_22_231229_create_units_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('product_id')->references('id')->on('products');
             $table->string('name');
-            $table->decimal('conversion_factor', 10, 4);
+            $table->float('conversion_factor');
             $table->timestamps();
             $table->softDeletes();
         });
